Refactor Loader icon to use a div wrapper
This change wraps the SVG loader icon in a div element. This allows the
icon to properly inherit the `class` and other props from its parent,
which is necessary for the `animate-spin` class to function correctly. A
`ts-expect-error` comment is added to suppress a TypeScript warning
related to SVG not being a composable transform target.